chore: refactor functions within entrypoints/content/ (#14)
