+++ src/entrypoints/content/comparison-popup-content.tsx
